Created mir_to_goto folder, moved monomorphize into it
