+++ compiler/rustc_codegen_llvm/src/gotoc/backend.rs
@@ -3,11 +3,11 @@
 
 //! This file contains the code necessary to interface with the compiler backend
 
-use super::cbmc::goto_program::symtab_transformer;
-use super::cbmc::goto_program::SymbolTable;
-use super::cbmc::{MachineModel, RoundingMode};
-use super::metadata::*;
-use super::monomorphize;
+use crate::gotoc::cbmc::goto_program::symtab_transformer;
+use crate::gotoc::cbmc::goto_program::SymbolTable;
+use crate::gotoc::cbmc::{MachineModel, RoundingMode};
+use crate::gotoc::metadata::*;
+use crate::gotoc::mir_to_goto::monomorphize;
 use bitflags::_core::any::Any;
 use rustc_codegen_ssa::traits::CodegenBackend;
 use rustc_data_structures::fx::FxHashMap;
