+++ compiler/rustc_codegen_llvm/src/gotoc/mir_to_goto/mod.rs
@@ -1,0 +1,5 @@
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0 OR MIT
+
+pub mod monomorphize;
+//pub mod stubs;
